@@ -1,6 +1,6 @@
 module github.com/mrlyc/cmdr
 
-go 1.22
+go 1.24
 
 require (
 	github.com/ahmetb/go-linq/v3 v3.2.0
